pyproject.toml
[tool.poetry]
name = "snowbear"
version = "0.2.7"
description = ""
authors = ["Omri Fima <omrif@diagnosticrobotics.com>"]

[tool.poetry.dependencies]
python = ">=3.7.4,<4.0"
snowflake-connector-python = ">=2.7.0"
pandas = ">=1.3.0"
PyPika = "^0.48.9"
snowflake-sqlalchemy = ">=1.3.0"
pyarrow = ">=5.0.0,<6.1.0"

[tool.poetry.dev-dependencies]
pytest = "7.1.1"
isort = "^5.7.0"
deepdiff = "^5.2.3"
flake8 = "^4.0.1"
flake8-tidy-imports = "^4.5.0"
pycodestyle = "^2.8.0"
pyflakes = "^2.4.0"
flake8-pytest-style = "^1.5.1"
flake8-quotes = "^3.3.1"
pydocstyle = "^6.1.1"
black = "^22.6.0"



[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

